add measurement result response
