+++ Assets/Dataskop/Scripts/Core/Data/Models/MeasurementResultResponse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eaededfee826c7148a655546bd6add4c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
